+++ vscode/src/utils/FileSystem.ts
@@ -42,15 +42,18 @@ class NodeFileSystem implements FileSystem {
   };
 }
 
-/** Returns path2 relative to path1 */
-export const relativePath = (path1: string, path2: string) => {
+/**
+ * Returns path2 relative to path1, in posix format on all platforms.
+ * Input paths can be windows paths.
+ */
+export const posixRelativePath = (path1: string, path2: string) => {
   if (_path.extname(path1) !== '') {
     // ensure path 1 is a dir, otherwise path.relative doesn't work
     path1 = _path.dirname(path1);
   }
   let relPath = _path
     .relative(path1, path2)
-    .replace('\\', '/');
+    .replace(/\\/g, '/');
   if (relPath.startsWith('/')) { relPath = relPath.slice(1, relPath.length); }
   return relPath;
 };